! Start to pull theme JS and CSS in to their own class cleans up theme.php which was getting to an overwhelming number of methods also any hardening and easy cleanup === [] etc
